refactor: `src/components/panel`의 컴포넌트들을 `src/components/home`으로 이동한다
