+++ src/components/home/InfinitePanelGrid.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 
-import PanelItem from '@/components/panel/PanelItem';
+import PanelItem from '@/components/home/PanelItem';
 import { useMyPanelsInfiniteQuery } from '@/hooks/panel';
 import useInView from '@/hooks/useInView';
