+++ src/components/home/PanelItem.tsx
